backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
setuptools==70.1.1
mysqlclient==2.2.4
django-cors-headers==4.4.0
django-environ==0.11.2